favList

+++ src/API/modules/AdminBooking.ts
@@ -1,15 +1,15 @@
 import axiosClient from "../axsiosClient";
 
 export const getAllBookings = () => {
-    return axiosClient.get("/booking");
+    return axiosClient.get("/api/v0/admin/booking");
 };
 
 
 export const viewBookings = (id: number) => {
-    return axiosClient.get(`/booking/${id}`);
+    return axiosClient.get(`/api/v0/admin/booking/${id}`);
 };
 
 export const deleteBookings = (id: number) => {
-    return axiosClient.delete(`/booking/${id}`);
+    return axiosClient.delete(`/api/v0/admin/booking/${id}`);
 };
